Business Case:  FicZon Inc is an IT solution provider with products ranging from on premises products to SAAS based solutions.
FicZon major leads  generation channel is digital and through their website.   FicZon business is majorly dependent on the sales force  effectiveness. As the market is maturing and more new competitors  entering the market, FicZon is experiencing the dip in sales.   Effective sales is dependent on lead quality and as of now, this is  based on manual categorization and highly depended on sales staff.  Though there is a quality process, which continuously updates the  lead categorization, it’s value is in for post analysis, rather than  conversation.  FicZon wants to explore Machine Learning to pre-categorize the lead  quality and as result, expecting significant increase in sales  effectiveness.  PROJECT GOAL:  1. Data exploration insights – Sales effectiveness.  2. ML model to predict the Lead Category (High Potential , Low  Potential)